+++ lua/projector/config.lua
@@ -2,7 +2,7 @@ local config = {
   loaders = {
     {
       module = require 'projector.loaders.legacy',
-      path = "/home/andrej/Repos/nvim-projector/examples/projector.json",
+      path = vim.fn.getcwd() .. '/.vscode/projector.json',
     }
   },
   outputs = {
